+++ catkin_ws/src/Task2/scripts/decreasing.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int64
+from std_msgs.msg import Int32
 
 
 dec = 0
 rospy.init_node("decreasing")
 rate = rospy.Rate(1)
-publisher = rospy.Publisher("/dec", Int64)
+publisher = rospy.Publisher("/dec", Int32)
 
 while not rospy.is_shutdown():
     publisher.publish(dec)
