+++ catkin_ws/src/Task2/scripts/increasing.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int64
+from std_msgs.msg import Int32
 
 
 inc = 0
 rospy.init_node("increasing")
 rate = rospy.Rate(1)
-publisher = rospy.Publisher("/inc", Int64)
+publisher = rospy.Publisher("/inc", Int32)
 
 while not rospy.is_shutdown():
     publisher.publish(inc)
